+++ strength_design/lib/src/data/models/exercise.dart
@@ -1,4 +1,3 @@
-
 class Exercise {
   final String id;
   final String name;