+++ strength_design/lib/src/data/models/food.dart
@@ -1,0 +1,19 @@
+class Food {
+  final String barcode;
+  final String name;
+  final String brand;
+  final double calories;
+  final double protein;
+  final double carbs;
+  final double fat;
+
+  Food({
+    required this.barcode,
+    required this.name,
+    required this.brand,
+    required this.calories,
+    required this.protein,
+    required this.carbs,
+    required this.fat,
+  });
+}