+++ strength_design/lib/src/data/repositories/exercise_repository.dart
@@ -1,4 +1,3 @@
-
 import 'package:cloud_firestore/cloud_firestore.dart';
 import 'package:sqflite/sqflite.dart';
 import 'package:strength_design/src/data/models/exercise.dart';
@@ -14,13 +13,16 @@ class ExerciseRepository {
   Future<List<Exercise>> getExercises() async {
     try {
       // First, try to fetch from the local database
-      final List<Map<String, dynamic>> maps = await _database.query('exercises');
+      final List<Map<String, dynamic>> maps = await _database.query(
+        'exercises',
+      );
       if (maps.isNotEmpty) {
         return maps.map((map) => Exercise.fromMap(map)).toList();
       }
 
       // If the local database is empty, fetch from Firestore
-      final QuerySnapshot snapshot = await _firestore.collection('exercises').get();
+      final QuerySnapshot snapshot =
+          await _firestore.collection('exercises').get();
       final List<Exercise> exercises = snapshot.docs.map((doc) {
         return Exercise.fromMap(doc.data() as Map<String, dynamic>);
       }).toList();